# [perf] AssetKey.__eq__ tweaks