@alangenfeld 2024-05-01
> very curious about the historical context of the previous implementation

https://github.com/dagster-io/dagster/pull/10139